Extract config dialogs into WindowConfigDialogsMixin
Continue decomposing window.py: move the known-hosts editor launcher, the
preferences window launcher, and the config export/import flow (export dialog,
import dialog, import-mode prompt, and the import itself) verbatim into
sshpilot/window_dialogs.py as WindowConfigDialogsMixin, inherited by MainWindow.
window.py shrinks ~305 lines.

Pure code move; the module-level names the block relied on (PreferencesWindow,
get_config_dir, datetime, json) are imported into the mixin, while BackupManager
/ KnownHostsEditorWindow / icon_utils stay as local imports. The generic
_error_dialog / _info_dialog helpers these call remain in window.py and resolve
via self.

Also drop a now-redundant local `from datetime import datetime` in
on_export_diagnostics_action (the module-level import covers it; removing the
shadow clears an F811 surfaced by the move).

Add tests/test_window_config_dialogs_mixin.py asserting the methods resolve to
the mixin module and the mixin is in the MRO.
